Add separate file for `DestroyContext` command.
